requirements.txt
PyQt6>=6.7
qrcode[pil]>=7.4
Pillow>=10.0
cryptography>=46.0.7
requests>=2.31
PyYAML>=6.0
